requirements.txt
albucore==0.0.13
albumentations==1.4.10
annotated-types==0.7.0
anyio==4.6.2.post1
astor==0.8.1
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
fire==0.7.0
fonttools==4.55.0
google-api-core==2.23.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.36.0
imgaug==0.4.0
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lmdb==1.5.1
lxml==5.3.0
matplotlib==3.9.2
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.9.1
paddlepaddle==2.6.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
PyYAML==6.0.2
RapidFuzz==3.10.1
requests==2.32.3
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.5.0
shapely==2.0.6
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tomli==2.1.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
